# src/backend/CompanyHero.Api/Contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CompanyHero API",
    "description": "Verbindlicher Vertrag nach A-009 und K13: Kennungen und Dezimalwerte als Strings, Zeitpunkte mit Offset, Null und fehlendes Feld unterscheidbar.",
    "version": "1"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/members": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ListMembers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/auth/providers": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ListLoginProviders",
        "parameters": [
          {
            "name": "tenant",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/oidc/{key}/start": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "StartExternalLogin",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "intent",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/auth/passkey": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "CompletePasskeyLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/recovery": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "RecoveryCodeLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/magic-link": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "RequestMagicLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/auth/magic-link/consume": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ConsumeMagicLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/transfer": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ConsumeTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JsonElement": {},
...
      "LoginResponse": {
        "required": [
          "tenantId",
          "personId",
          "kind",
          "newRecoveryCode",
          "mustSetUpAccess"
        ],
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "personId": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SessionKindDto"
          },
          "newRecoveryCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "mustSetUpAccess": {
            "type": "boolean"
          }
        }
      },
      "MagicLinkRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "MeResponse": {
        "required": [
          "personId",
          "displayName",
          "tenantId",
          "tenantName",
          "roles"
        ],
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MemberResponse": {
        "required": [
          "personId",
          "displayName",
          "roles",
          "joinedAt"
        ],
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PasskeyAnswerRequest": {
        "required": [
          "state",
          "credential",
          "deviceName"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "credential": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "deviceName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProviderResponse": {
        "required": [
          "key",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "RecoveryLoginRequest": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
...
      "SessionKindDto": {
        "enum": [
          "member",
          "privileged",
          "kiosk_device",
          "kiosk_person"
        ]
      },
...
      "TokenRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}